# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Receipt Assistant API",
    "version": "2.0.0-alpha",
    "description": "Double-entry ledger backend. The v1 resources (transactions, postings, accounts, documents) are under active construction — see issues #33 (schema), #35 (transactions API), #36 (accounts API)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "CreateTransactionRequest": {
        "type": "object",
        "properties": {
          "occurred_on": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2026-04-19"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time"
          },
          "payee": {
            "type": "string"
          },
          "narration": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "posted",
              "voided",
              "reconciled",
              "error"
            ]
          },
          "postings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewPosting"
            },
            "minItems": 2
          },
          "document_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            }
          },
          "trip_id": {
            "type": "string",
            "format": "uuid",
            "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {},
            "default": {},
            "description": "User-defined JSON object; not schema-validated."
          }
        },
        "required": [
          "occurred_on",
          "postings"
        ]
      },
...
      "VoidTransactionRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/v1/transactions": {
      "post": {
        "summary": "Create a transaction",
        "tags": [
          "transactions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Idempotency-Key required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List transactions",
        "tags": [
          "transactions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2026-04-19"
            },
            "required": false,
            "name": "occurred_from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2026-04-19"
            },
            "required": false,
            "name": "occurred_to",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "amount_min_minor",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ]
            },
            "required": false,
            "name": "amount_max_minor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": false,
            "name": "account_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "payee_contains",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "posted",
                "voided",
                "reconciled",
                "error"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": false,
            "name": "trip_id",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            },
            "required": false,
            "name": "has_document",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": false,
            "name": "source_ingest_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "occurred_on",
                "amount",
                "created_at"
              ]
            },
            "required": false,
            "name": "sort",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "required": false,
            "name": "order",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      }
                    },
                    "next_cursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "next_cursor"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/{id}": {
      "get": {
        "summary": "Get a transaction",
        "tags": [
          "transactions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch transaction head fields",
        "tags": [
          "transactions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "412": {
            "description": "Version mismatch",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "If-Match required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete draft/error transaction",
        "tags": [
          "transactions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "409": {
            "description": "Must void instead",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/{id}/void": {
      "post": {
        "summary": "Void a posted transaction",
        "tags": [
          "transactions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mirror transaction created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/{id}/reconcile": {
      "post": {
        "summary": "Reconcile a posted transaction",
        "tags": [
          "transactions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/bulk": {
      "post": {
        "summary": "Bulk update/void/reconcile",
        "tags": [
          "transactions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Per-op results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/{id}/postings": {
      "post": {
        "summary": "Add a posting to a transaction",
        "tags": [
          "transactions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "01HXY9F0ABCDEFGHJKMNPQRSTV"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "If-Match",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPosting"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Posting added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Posting"
                }
              }
            }
          },
          "422": {
            "description": "Imbalance",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}